refactor: 쿠폰 등록 api 함수 파일 위치 이동
